Implementação do Sicredi na nova estrutura
